Fixed Warnings
Fixed:
File Helper::checkNextChar
Warning	C4267	'initializing': conversion from 'size_t' to 'int', possible loss of data	
BoardSquareInfo::SetDisplayName
Warning	C4267	'initializing': conversion from 'size_t' to 'int', possible loss of data
Piece::getCharFromMyPiece
Warning	C4715	'Piece::getCharFromMyPiece': not all control paths return a value

+++ RockPaperScissors/BoardSquareInfo.cpp
@@ -17,7 +17,7 @@ BoardSquareInfo::BoardSquareInfo(BoardPoint i_PointInfo, const char * i_DisplayN
 void BoardSquareInfo::SetDisplayName(const char * i_DisplayName)
 {
 	strcpy_s(displayName, i_DisplayName);
-	int strEndIndex = strlen(i_DisplayName);
+	size_t strEndIndex = strlen(i_DisplayName);
 	displayName[strEndIndex] = '\0';
 }
